requirements.txt
unoserver
python-pptx
Cython
opencv-python
numpy
ffmpeg-python
psutil
pyopenjtalk
espnet
espnet-model-zoo
soundfile
pydub
moviepy

torch==1.11.0
torchvision==0.12.0
torchaudio==0.11.0
torchtext==0.12.0
